Rename Client::System::Framework::ClipboardViewerHandler to Client::UI::UIClipboard
